src/python/requirements.txt
zensols.db~=1.3.1
zensols.deepnlp~=1.13.0
